+++ public/index.html
@@ -81,12 +81,12 @@
       <option value="Willis">
       <option value="Wallace">
     </datalist>
-    <script type="text/javascript" src="ws.js"></script>
     <script type="text/javascript" src="update.js"></script>
     <script type="text/javascript" src="insert.js"></script>
     <script type="text/javascript" src="wsout.js"></script>
-    <script type="text/javascript" src="wswait.js"></script>
+    <script type="text/javascript" src="ui.js"></script>
     <script type="text/javascript" src="main.js"></script>
+    
 </body>
 
 </html>
