+++ public/wsout.js
@@ -1,6 +1,26 @@
 /* global pNameToElm*/
+window.WS = {};
+window.WS.makews = function makews() {
+  let ws = new WebSocket(
+    "wss://" + window.location.hostname + window.location.pathname
+  );
 
-window.WS.ws.onmessage = (ev) => {
+  ws.onopen = () => {
+    console.log("[WS Open]")
+    if (!window._ps) {
+      //Cries in development
+      if ( window.location === window.parent.location ) {window._ps = prompt("Password");}
+    }
+    ws.send(JSON.stringify({ type: "Auth", pass: window._ps + "" }));
+  };
+  window.WS.ws = ws;
+  return ws;
+};
+window.WS.makews();
+window.WS.send = function send(obj,type="update") {
+  window.WS.ws.send(JSON.stringify({ type: type, data: obj }));
+};
+function onmessage(ev) {
   if (ev.data === "connect") {
     return;
   }
@@ -19,8 +39,14 @@ window.WS.ws.onmessage = (ev) => {
         .getElementsByClassName(data[1].classes)[0].value = data[1].value;
     }
   }
-};
-window.WS.ws.onclose = function (event) {
-  //this doesn't really work lol
+  if (type === "info") {
+  }
+}
+window.WS.ws.onmessage = onmessage;
+function close() {
+  console.log("[WS Close]");
   window.WS.ws = window.WS.makews();
-};
+  window.WS.ws.onmessage = onmessage;
+  window.WS.ws.onclose= close;
+}
+window.WS.ws.onclose = close
