+++ src/index.js
@@ -9,22 +9,6 @@ const WebSocket = require("ws");
 const websocketServer = new WebSocket.Server({ server });
 const wsswrap = require("./wsswrap");
 wsswrap(websocketServer);
-//when a websocket connection is established
-/*
-websocketServer.on("connection", (webSocketClient) => {
-  //send feedback to the incoming connection
-  webSocketClient.send('{ "connection" : "ok"}');
-
-  //when a message is received
-  webSocketClient.on("message", (message) => {
-    //for each websocket client
-    websocketServer.clients.forEach((client) => {
-      //send the client the current message
-      client.send(`{ "message" : ${message} }`);
-    });
-  });
-});*/
-
 //start the web server
 server.listen(process.env.PORT || serverPort, () => {
   console.log(`Websocket server started on port ` + serverPort);
